Add: only buffer at most 5 max-sized packets per client (#20)
This means that when a user downloads a 500+ MiB content entry,
and is on a slow connection, the asyncio write() task will no
longer cache this in full in the memory. Now it will "async" out
till there is room in the write buffer again before continueing
to write. This should heavily reduce the memory footprint in
production.

Additionally, raise an exception if the socket is being closed,
instead of returning False. This means it bubbles up for all
send_packet() calls, which should reduce the amount of warnings
we are now seeing about writing to a closed socket.