Add: an option to indicate the first entry in BaseGraphics listing (#22)
OpenTTD client's bootstrap implementation downloads what-ever is
the first entry of the BaseGraphics listing. Without this, this
strongly depends on the order of listdir(), but rarely is OpenGFX,
which we prefer. Of course this can be considered a bug in OpenTTD
client, but for now we can fix this by adding an option that forces
a predefined unique-id to be the first in the list. In production
this will of course be OpenGFX.