Fix: close the client-connection if the s3-backend-connection fails
Now it triggered an exception, crashing the server. This happens
rarely, and only if a client connection is very slow. As this is
already the fallback of the normal download flow, we are left
with no alternatives: the client cannot download the file via a
more reliable HTTP connection, and is downloading over our TCP
connection in such a slow speed, the S3 backend gives up on it.
The only reasonable thing to do, is to drop the connection to
the client on our end too.