Fix: crash with invalid version for a branch with CLIENT_INFO_LIST (#64)
There is a choice to make:
1) parse as much of the version as possible
2) ignore the branch
3) ignore the request

This commit goes for 3), as that is the most verbose to the developer
working on OpenTTD. It is a protocol violation, and the developer
should address the issue. With 1) and 2) it could remain invisible
for a very long time, till it breaks. 3) is the safest, as it
will be noticed sooner or later.