Feature: use classification as tags (#300)
This is mostly for backwards compatibility; as we no longer allow
custom tags, the best-next-thing are the classification markers.

This means in current OpenTTD clients they show up as tags, in
which you can search etc. This was kinda how tags were meant, but
sadly people decided to use them for completely different purposes.